feat(gestion productos y categoria): funcionalidad de agregar producto lista, endopinds de producto y categoria listos

+++ src/main/java/com/web/prime_drip_club/models/DetallePedido.java
@@ -12,9 +12,9 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class DetallePedido {
-    private Integer id;
-    private Integer pedidoId;
-    private Integer productoId;
+    private Long id;
+    private Long pedidoId;
+    private Long productoId;
     private Integer cantidad;
     private BigDecimal precioUnitario;
 }
